+++ service/src/main/java/bio/terra/landingzone/stairway/flight/create/resource/step/CreatePostgresqlDbStep.java
@@ -38,7 +38,10 @@ public class CreatePostgresqlDbStep extends BaseResourceCreateStep {
   public StepResult undoStep(FlightContext context) throws InterruptedException {
     var postgresId = context.getWorkingMap().get(POSTGRESQL_ID, String.class);
     try {
-      armManagers.postgreSqlManager().servers().deleteById(postgresId);
+      if (postgresId != null) {
+        armManagers.postgreSqlManager().servers().deleteById(postgresId);
+        logger.info("{} resource with id={} deleted.", getResourceType(), postgresId);
+      }
     } catch (ManagementException e) {
       if (StringUtils.equalsIgnoreCase(e.getValue().getCode(), "ResourceNotFound")) {
         return StepResult.getStepResultSuccess();
